Change the way we will test.
Due to Test::Testers lack of subtest testing support, and my laziness
to learn the Test2 framework, we will now test match/no match in
a different way.

We first run the tests in a child process, collecting the output (and
error/diagnostic messages). Then, in a separate step, we will inspect
the collected output, and see whether it outputs what it should output.

This way, the calls to 'ok', 'is', etc, done by match/no_match, will
not interfere with the actual tests run.